backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.40.0
httpx==0.27.0
faster-whisper==0.10.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
alembic==1.12.1
pytz==2024.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
pydantic[email]==2.5.0
bcrypt==4.0.1